Add the prompt message, asking user for the size
Initialize the variable 'size' with the value of prompt converted to a
number. If the 'size' is smaller than 0 or larger than 100, set the
'size' to 0 and display a message. This is needed, so that the user
cannot generate a grid large enough, for the browser to freeze or crash.

Edit the gridTemplateColumns and grid TemplateRows to use the ${size}
as their first argument. This is essential for the js code to run
properly, as it sets the number of rows and columnds to be equal the
value of the 'size' variable.

Edit the style.css to make all children of the container to not shrink.
Set the height and width of the container to be equal 960px, because it
is needed for the container to remain the same size, even with different
inputs from the user.