Implement a button that toggles random color
Create a random() function that initializes three values with the value
of a rounded (Math.random() * 255) for all three values and set the
return string to return all values in rgb(value1, value2, value3) format.

Add the button that changes the element's background color when hovering
on it to the value of a return from the function random().

Change the style.css in order to position the button beneath the grid.